refactor: 커스텀훅 중 도메인과 상관없이 범용적인 훅은 shared 폴더로 이동
커스텀훅 중 도메인을 알고 있는 것과 그렇지 않은 것을 분리하면 훅 사용 시 헷갈리지 않고 쓸 수 있을 것으로 기대
