+++ src/hooks/shared/usePost.js
@@ -1,13 +1,10 @@
 import { useState } from 'react';
 import apiClient from 'utils/apiClient';
-// import { useDispatch } from 'react-redux';
-// import { getCartListAsync } from 'reducers/cartList/cartList.thunks';
 
 const usePost = (url, payload) => {
   const [isLoading, setIsLoading] = useState(false);
   const [isError, setIsError] = useState(false);
   const [result, setResult] = useState(null);
-  // const dispatch = useDispatch();
 
   const callPostApi = () => {
     setIsLoading(true);
@@ -17,7 +14,6 @@ const usePost = (url, payload) => {
       .then((response) => {
         setIsLoading(false);
         setResult(response.data);
-        // dispatch(getCartListAsync); // api에서 리덕스로 cart 상태 가져오기
       })
       .catch((error) => {
         console.log(error);
